refactor: locator helper
