+++ lib/abi/IStakingModule.json
@@ -1,1 +1,1 @@
-[{"anonymous":false,"inputs":[{"indexed":false,"name":"validatorsKeysNonce","type":"uint256"}],"name":"ValidatorsKeysNonceChanged","type":"event"}]
+[{"constant":false,"inputs":[{"name":"nodeOperatorId","type":"uint256"},{"name":"stuckSigningKeysCount","type":"uint64"}],"name":"updateStuckValidatorsKeysCount","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"anonymous":false,"inputs":[{"indexed":false,"name":"validatorsKeysNonce","type":"uint256"}],"name":"ValidatorsKeysNonceChanged","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"nodeOperatorId","type":"uint256"},{"indexed":false,"name":"stuckValidatorsCount","type":"uint256"}],"name":"StuckValidatorsCountChanged","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"nodeOperatorId","type":"uint256"},{"indexed":false,"name":"forgivenValidatorsCount","type":"uint256"}],"name":"ForgivenValidatorsCountChanged","type":"event"}]
