+++ test/0.8.9/lido-locator.test.js
@@ -1,32 +1,15 @@
 const { ZERO_ADDRESS } = require('@aragon/contract-helpers-test')
-const { randomBytes } = require('crypto')
-const { toChecksumAddress } = require('ethereumjs-util')
 const hre = require('hardhat')
 const { assert } = require('../helpers/assert')
+const { locatorServices, getRandomLocatorConfig } = require('../helpers/locator')
 const Proxy = artifacts.require('OssifiableProxy.sol')
 const LidoLocator = artifacts.require('LidoLocator.sol')
-
-const services = [
-  'accountingOracle',
-  'depositSecurityModule',
-  'elRewardsVault',
-  'legacyOracle',
-  'lido',
-  'safetyNetsRegistry',
-  'selfOwnedStEthBurner',
-  'stakingRouter',
-  'treasury',
-  'validatorExitBus',
-  'withdrawalQueue',
-  'withdrawalVault',
-  'postTokenRebaseReceiver'
-]
 
 contract('LidoLocator', ([deployer, agent]) => {
   let evmSnapshotId
   let proxy
   let lidoLocatorProxy
-  const initialConfig = getRandomConfig()
+  const initialConfig = getRandomLocatorConfig()
 
   before(async () => {
     const implementation = await LidoLocator.new(initialConfig, { from: deployer })
@@ -75,10 +58,8 @@ contract('LidoLocator', ([deployer, agent]) => {
   describe('breaking constructor', () => {
     it('should revert when passing a zero address', async () => {
       const configsWithZeroAddress = []
-      for (const service of services) {
-        const config = getRandomConfig()
-        config[service] = ZERO_ADDRESS
-
+      for (const service of locatorServices) {
+        const config = getRandomLocatorConfig({ [service]: ZERO_ADDRESS })
         configsWithZeroAddress.push(config)
       }
 
@@ -90,7 +71,7 @@ contract('LidoLocator', ([deployer, agent]) => {
 
   describe('checking updated implementation', () => {
     describe('works after upgrade to a compatible impl', () => {
-      const updatedConfig = getRandomConfig()
+      const updatedConfig = getRandomLocatorConfig()
 
       beforeEach(async () => {
         const updatedImplementation = await LidoLocator.new(updatedConfig, { from: deployer })
@@ -129,14 +110,3 @@ contract('LidoLocator', ([deployer, agent]) => {
     })
   })
 })
-
-function getRandomConfig() {
-  return services.reduce((config, current) => {
-    config[current] = generateRandomAddress()
-    return config
-  }, {})
-}
-
-function generateRandomAddress() {
-  return toChecksumAddress('0x' + randomBytes(20).toString('hex'))
-}
